@philipaconrad 2022-08-08
I'm going to add some correctness/regression tests, at the suggestion of @charlesdaniels.

The tests will compare the behavior of `union(...)` against pairwise `Set.Union` operations for various lists of input sets. These should be cheap to run, and will help us prevent a silly regression if our existing test suite doesn't already check for this stuff.