@philipaconrad 2022-08-08
I've split the `union` tests/benchmarks off into their own files, since I'm planning to soon be doing the same process for the `intersection` builtin, and `topdown_bench_test` was getting a bit crowded.