@tsandall 2022-08-11
LGTM. In this case, inlining the union logic is cheap enough that the speedup is worth it.